### 2022-04-01 #4293 @kmongird
Hello @fraukewiese I have added the missing DOI to the reference mentioned above and my suggested reviewers from the JOSS reviewer list are: wiljnich and j3r3m1. Thanks!
