### 2022-04-20 #4293 @wiljnich
I am willing to review this submission.

> On Apr 20, 2022, at 6:09 AM, fraukewiese ***@***.***> wrote:
>
> ﻿
> @you , @wiljnich , @j3r3m1 - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
