### 2022-04-22 #4336 @editorialbot
[REVIEW]: interflow: A Python package to organize, calculate, and visualize sectoral interdependency flow data
**Submitting author:** @kmongird (<a href="http://orcid.org/0000-0003-2807-7088">Kendall Mongird</a>)
**Repository:** https://github.com/pnnl/interflow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
