### 2022-04-22 #4336 @wiljnich
I am very pleased with this package and happily recommend interflow for publication in JOSS. I have one recommendation, which is not acceptance blocking. To conform with the review criteria (Documentation #5), I believe that a short explainer of the test cases and instructions for their use should be added to the documentation site.

This is a great package that solves a painful need - I have built Sankey energy flows in Python before, and this is a vast improvement. Excellent work by @kmongird and team.
