### 2022-04-22 #4336 @kmongird
@wiljnich thank you very much for your review! We greatly appreciate your time and comments.

Just to get some clarification on the recommendation you’ve described and to make sure we’re on the same page, when you state “short explainer of the test cases and instructions for their use should be added” are you referring to the test suite for the package or the usage of the sample data in the quickstarter?
