### 2022-05-11 pnnl/interflow#8 @j3r3m1
JOSS review: Software paper organization
Refering to JOSS paper (https://github.com/openjournals/joss-reviews/issues/4336), I suggest that you slightly reorganize the text to fill all sections requested by the template:

**A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?**
-> you do not have this section. You have most of the content in the summary but I think it can be moved (or repeated with a longer description) in this section

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**
-> You actually write this section under the wrong title ("A statement of the need"). I suggest you move things to this section. I have also found a couple of tools (free and open source or not...) allowing [to do Sankey Diagrams](https://www.sankey-diagrams.com/sankey-diagram-software/). It might be of interest to discuss the main differences with some of them if you feel it is needed.

**Other potential section:**
I think adding a chart showing a workflow of what is done by "interflow" when applying one of the main function could draw much more users.
