### 2022-05-11 pnnl/interflow#9 @j3r3m1
[BUG] Trouble with the input data zip file ?
Refering to JOSS paper (https://github.com/openjournals/joss-reviews/issues/4336).

By curiosity I have downloaded the zip file [here](https://github.com/pnnl/interflow/blob/main/interflow/input_data/us_county_sample_data.csv.zip) but there is a bug if I try to extract the data manually. Is this a bug ?
