### 2022-05-11 pnnl/interflow#10 @j3r3m1
Adding a more simple case from scratch ?
Refering to JOSS paper (https://github.com/openjournals/joss-reviews/issues/4336), I would be glad to be able to produce (at least some of the analysis) from scratch since in all examples I have found you only use the US data.

More precisely, and I think it is related to my demand about a chart in #8, it would be nice to have:
- either an example in your jutyter notebook creating and calculating a very limited dataset
- either a chart summarizing the format of the needed inputs in the documentation or in the paper (or in both). There is a description [here](https://pnnl.github.io/interflow/user_guide.html#creating-input-data-for-different-calculations) but it is a quite long description and I am not sure about the number of DataFrame needed as input neither totally sure about their content.
