+++ interflow/sample_data.py
@@ -176,9 +176,6 @@ def calc_irrigation_consumption() -> pd.DataFrame:
         df['IC_CU_RWW_frac'] = np.where(df['State'] == state, df['IR_CU_RWW_frac'],
                                         df['IC_CU_RWW_frac'])  # reclaimed
 
-    # preset interbasin transfer to crop irrigation consumption equal to fresh surface water consumption
-    df['IC_CU_ibt_frac'] = df['IC_CU_FSW_frac']
-
     # rename variables
     variable_dict = {'FIPS': 'FIPS',
                      'State': 'State',
@@ -188,8 +185,7 @@ def calc_irrigation_consumption() -> pd.DataFrame:
                      'IC_CU_RWW_frac': 'AGR_crop_reclaimed_wastewater_import_mgd',
                      'IG_CU_FSW_frac': 'AGR_golf_fresh_surfacewater_withdrawal_mgd',
                      'IG_CU_FGW_frac': 'AGR_golf_fresh_groundwater_withdrawal_mgd',
-                     'IG_CU_RWW_frac': 'AGR_golf_reclaimed_wastewater_import_mgd',
-                     'IC_CU_ibt_frac': 'AGR_crop_ibt_total_import_mgd'}
+                     'IG_CU_RWW_frac': 'AGR_golf_reclaimed_wastewater_import_mgd'}
     variable_list = list(variable_dict.keys())
     df = df[variable_list]
     df = df.rename(columns=variable_dict)
@@ -197,8 +193,7 @@ def calc_irrigation_consumption() -> pd.DataFrame:
     # create a list of sector water withdrawal variable name starters
     flow_list = ['AGR_crop_fresh_surfacewater_withdrawal_mgd', 'AGR_crop_fresh_groundwater_withdrawal_mgd',
                  'AGR_crop_reclaimed_wastewater_import_mgd', 'AGR_golf_fresh_surfacewater_withdrawal_mgd',
-                 'AGR_golf_fresh_groundwater_withdrawal_mgd', 'AGR_golf_reclaimed_wastewater_import_mgd',
-                 'AGR_crop_ibt_total_import_mgd']
+                 'AGR_golf_fresh_groundwater_withdrawal_mgd', 'AGR_golf_reclaimed_wastewater_import_mgd']
 
     # create a consumption name adder to add on to variable names
     adder = '_to_CMP_total_total_total_total_mgd_fraction'
@@ -399,7 +394,6 @@ def calc_irrigation_conveyance_loss_fraction(loss_cap=True, loss_cap_amt=.90) ->
     created_variable_list = ['AGR_crop_fresh_surfacewater_withdrawal_mgd',
                              'AGR_crop_fresh_groundwater_withdrawal_mgd',
                              'AGR_crop_reclaimed_wastewater_import_mgd',
-                             'AGR_crop_ibt_total_import_mgd',
                              'AGR_golf_fresh_surfacewater_withdrawal_mgd',
                              'AGR_golf_fresh_groundwater_withdrawal_mgd',
                              'AGR_golf_reclaimed_wastewater_import_mgd']
@@ -431,7 +425,6 @@ def calc_irrigation_discharge_flows():
     variable_list = ['AGR_crop_fresh_surfacewater_withdrawal_mgd',
                      'AGR_crop_fresh_groundwater_withdrawal_mgd',
                      'AGR_crop_reclaimed_wastewater_import_mgd',
-                     'AGR_crop_ibt_total_import_mgd',
[…]
