+++ paper/paper.md
@@ -37,9 +37,9 @@ The `interflow` package provides a flexible tool to organize, calculate, and vis
 
 The ability to calculate and organize sectoral interdependency data and evaluate where the greatest cross-sectoral intensities and flows exist can enhance network resilience and understanding across both sectors and regions. Historically, however, sectoral interdependency data can be difficult to find and complex to analyze, understand, and utilize. There is also significant variance in the availability of comprehensive and reproducible data sets, making it more difficult to build on prior research and make new advances. An open-source framework that can generate, organize, and visualize multisectoral interdependency data in a replicable manner would be advantageous to several audiences. This list can include researchers in energy, climate, water, agriculture, or other areas studying sectoral interactions as well as policymakers and regulators seeking information to better inform their decision-making. A generalizable tool that can identify where the interdependencies between sectors are, what their relative scale is, and how they compare across regions would contribute to advancement in several analytical spaces.
 
-# Statement of Need
+# State of the Field
 
-To the best of our knowledge, there is no open-source software option available that facilitates the calculation, organization, and visualization of multiscale, multisector interdependencies in one place. Software exists for visualizing pre-calculated sectoral flow values as Sankey diagrams such as the proprietary e!Sankey software [@esankey2022] or the Python library matplotlib [@hunter2007], but these tools do not offer a way to calculate the demand of a cross-resource type and build out new sectoral connections based on the result. The `interflow` package aims to fill this gap and serve as a flexible and open-source option for conducting multi-resource sectoral interdependency data calculations, producing multisectoral datasets, and visualizing the results.
+To the best of our knowledge, there is no open-source software option available that facilitates the calculation, organization, and visualization of multiscale, multisector interdependencies in one place. Software exists for visualizing pre-calculated sectoral flow values as Sankey diagrams such as e!Sankey [@esankey2022], SDraw [@sdraw2022], SankeyBuilder [@sankeybuilder2022], SankeyMATIC [@sankeymatic2022], or the Python library matplotlib [@hunter2007], but these tools do not offer a way to calculate the demand of a cross-resource type and build out new sectoral connections based on the result. While the proprietary options in the above list (e!Sankey, SDraw, and SankeyBuilder) allow for complex customizability in diagram formatting and, in the case of e!Sankey, can also calculate input and output flow balances by node, none of the options listed (proprietary or open-source) offer a cross-unit computation capability. The interflow package aims to fill this gap and serve as a flexible and open-source option for conducting multi-resource sectoral interdependency data calculations, producing multisectoral datasets, and visualizing the results.
 
 Sectoral interdependency analysis itself is not a new area of research [@curmi2013; @liu2016; @greenberg2017], but the publications in this area typically return only the end-product of their analysis (e.g., output values or diagrams). The detailed methodology or algorithms developed to conduct their calculations are often not provided in an accessible or transparent manner. Researchers looking to build upon or modify the existing assumptions and data are left to redevelop the calculation structure from scratch as a result. The `interflow` package provides a consistent and open-source calculation framework to enable reproducibility.
 
