+++ paper/paper.md
@@ -45,7 +45,9 @@ Sectoral interdependency analysis itself is not a new area of research [@curmi20
 
 # Design and Methodology
 
-The `interflow` package iterates through a user-provided tabular input data to (1) collect known resource flow values (e.g., water demand) between provided sectors, (2) calculate new sector flow values for a secondary resource type from the initial flow (e.g., energy demand based on the water flow) using provided cross-resource intensity coefficients, and (3) build upstream and downstream sector connections to carry those calculated and collected flows. Collected and calculated flows are then compiled in a way that can be used in a variety of data visualization functions offered in the package.
+The `interflow` package iterates through a user-provided tabular input data to (1) collect known resource flow values (e.g., water demand) between provided sectors, (2) calculate new sector flow values for a secondary resource type from the initial flow (e.g., energy demand based on the water flow) using provided cross-resource intensity coefficients, and (3) build upstream and downstream sector connections to carry those calculated and collected flows. Collected and calculated flows are then compiled in a way that can be used in a variety of data visualization functions offered in the package (Figure 2). 
+
+![Simplified Workflow Diagram of Collection and Calculation of Sectoral Flow Data.\label{fig:fig2}](fig2.png)
 
 The basic methodology described above is repeated for all regions provided by the user in the input data. Though the `interflow` package comes equipped with extensive sample data to evaluate U.S. county-level water and energy flows across various sectors, it can conduct analysis for any region (e.g., country, province), any sector or group of sectors (e.g., electricity generation, agriculture), and any group of resources (e.g., water, energy, food, carbon, land) and is limited only by the input data that the user provides. More information on the input data requirements can be found in the [generalizability documentation](https://pnnl.github.io/interflow/user_guide.html#generalizability). All code used to generate the sample water and energy input data for all US counties is additionally included as part of the package.
 
