### 2022-05-13 2b9a3d0
add wind consumption MSN codes for COM and IND
