### 2022-05-13 bb432ad
fix mining-other flow naming
