### 2022-05-13 fbfcfcb
adjust natgas and petroleum scaler
