+++ interflow/sample_data.py
@@ -2369,13 +2369,14 @@ def prep_pumping_energy_fuel_data() -> pd.DataFrame:
     # electricity source name
     egd_name = '_from_EGD_total_total_total_total_bbtu_fraction'
 
-    # list of non-irrigation sectors which are assumed to only take electricity
+    # list of non-irrigation or golf sectors which are assumed to only take electricity
     non_irr_sector_list = ['AGR_aquaculture_pumping_fresh_groundwater_bbtu',
                            'AGR_aquaculture_pumping_fresh_surfacewater_bbtu',
                            'AGR_aquaculture_pumping_saline_groundwater_bbtu',
                            'AGR_aquaculture_pumping_saline_surfacewater_bbtu',
                            'AGR_livestock_pumping_fresh_groundwater_bbtu',
-                           'AGR_livestock_pumping_fresh_surfacewater_bbtu']
+                           'AGR_livestock_pumping_fresh_surfacewater_bbtu',
+                           'AGR_golf_pumping_reclaimed_wastewater_bbtu']
 
     # loop through non-irrigation sectors and build full variable names and values
     for sector in non_irr_sector_list:
@@ -2493,7 +2494,8 @@ def prep_pumping_intensity_data() -> pd.DataFrame:
                'AGR_livestock_pumping_fresh_surfacewater_bbtu',
                'AGR_ethanol_pumping_fresh_surfacewater_bbtu',
                'PWS_pumping_fresh_surfacewater_total_bbtu',
-               'PWS_pumping_saline_surfacewater_total_bbtu']
+               'PWS_pumping_saline_surfacewater_total_bbtu',
+               'AGR_golf_pumping_reclaimed_wastewater_bbtu']
 
     for sw in sw_list:
         df[sw] = df['surface_water_pumping_bbtu_per_mg']
