### 2022-05-24 pnnl/interflow#8 @j3r3m1
Thank you for answering all points.
I am satisfied with all points except the one. The diagram shown Fig. 2 is quite complicated and it is not easy to relate it to the text you added in this section. I think it might be simplified to better show the link with the text (you have numbers in the text which can not be found in the figure for example).
