### 2022-05-24 #4336 @j3r3m1
All points have been quite well adressed, thank you to authors. However, the added diagram (Figure 2) is quite complicated to understand and hardly fit with text description. I have added a comment in the initial related issue (https://github.com/pnnl/interflow/issues/8).
