### 2022-05-26 #4336 @kmongird
@j3r3m1 thank you for the follow-up comments, we appreciate your thorough review. We have simplified the figure in the paper considerably and better aligned it with the text.
