### 2022-06-03 680f1bb
update variable naming in test
