+++ paper/paper.bib
@@ -75,7 +75,7 @@
 
   @techreport{greenberg2017,
     title={Development of Energy-Water Nexus State-level Hybrid Sankey Diagrams for 2010},
-    url={https://flowcharts.llnl.gov/content/assets/docs/2010_United-States_EnergyWater.pdf},
+    url={https://flowcharts.llnl.gov/report},
     author={Greenberg, Harris R. and Simon, A.J. and Singer, Suzanne L. and Shuster, Erik P.},
     year={2017},
     howpublished={[PDF]},
