+++ paper/paper.bib
@@ -1,10 +1,10 @@
 @inbook{webber2017,
-  	url = {https://doi.org/10.1016/B978-0-12-803237-4.00012-4},
   	author = {Webber, M.E.},
     chapter={2.2.6},
   	booktitle = {Competition for Water Resources: Experiences and Management Approaches in the US and Europe},
     ISBN = {9780128032374},
   	publisher = {Elsevier},
+  	doi = {10.1016/B978-0-12-803237-4.00012-4},
   	title = {{Chapter 2.2.6 - Water for Electricity Generation in the United States}},
   	year = 2017
 }
