### 2022-06-07 #4336 @kmongird
version tag for tagged release:
v1.0.3

Zenodo DOI:
10.5281/zenodo.6620928
