### 2022-06-08 #4336 @kmongird
Hi @fraukewiese,

On lines 103-104, the broken Greenberg et al. 2017 link has been removed and has been replaced with this working link:  https://flowcharts.llnl.gov/report

For the Webber 2017 reference, I removed the title link and have included the doi that is now shown on line 129: https://doi.org/10.1016/B978-0-12-803237-4.00012-4
